+++ nodes/__init__.py
@@ -1,11 +1,12 @@
 """ Node classes used by the Notification Node Server. """
 
-NSVERSION = "3.6.24"
+NSVERSION = "3.6.25"
 
 from .Notify import Notify
 from .Pushover import Pushover
 from .ISYPortal import ISYPortal
 from .TelegramUB import TelegramUB
+from .WhatsApp import WhatsApp
 from .AssistantRelay import AssistantRelay
 from .UDMobile import UDMobile
 from .Controller import Controller